assert mysterieux qui ne passe pas

+++ src/wattwatt/connectors/energyproviders/occasional/enginegenerator/EngineGeneratorConnector.java
@@ -31,7 +31,7 @@ public class EngineGeneratorConnector extends EnergyProviderConnector implements
 	}
 
 	@Override
-	public int fuelQuantity() throws Exception {
+	public double fuelQuantity() throws Exception {
 		return ((IEngineGenerator) this.offering).fuelQuantity();
 	}
